AGR-2845 refactored code changes
